backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
PyMuPDF==1.24.3
google-cloud-vision==3.7.2
web3==6.18.0
aiosqlite==0.20.0
python-Levenshtein==0.25.1
python-dotenv==1.0.1
filetype==1.2.0
pytest==8.2.0
pytest-mock==3.14.0
httpx==0.27.0
